Fix: replace parseInt() to parseFloat()

+++ js/style.js
@@ -9,27 +9,27 @@ const calculator = {
 
 function add(num1, num2)
 {
-    return parseInt(num1) + parseInt(num2);
+    return parseFloat(num1) + parseFloat(num2);
 }
 
 function subtract(num1, num2)
 {
-    return parseInt(num1) - parseInt(num2);
+    return parseFloat(num1) - parseFloat(num2);
 }
 
 function multiply(num1, num2)
 {
-    return parseInt(num1) * parseInt(num2);
+    return parseFloat(num1) * parseFloat(num2);
 }
 
 function divide(num1, num2)
 {
-    return parseInt(num1) / parseInt(num2);
+    return parseFloat(num1) / parseFloat(num2);
 }
 
 function pow(num1, num2)
 {
-    return parseInt(num1) ** parseInt(num2);
+    return parseFloat(num1) ** parseInt(num2);
 }
 
 function setNumber(value)
